refactor(ar/ml): finish Phase 3 audit — runOnce split, predictor injection, god-file extraction (#188)
P2.3 — runOnce decomposition
 * Split the 130-line `runOnce` into five named helpers : `runInference`,
   `fitInverseScale`, `fuseAndAccumulate`, `publish`. The top-level
   function is now a four-step pipeline that reads top-to-bottom.

P3.2 — predictor protocol injection
 * Added `SemSegPredicting` / `DepthPredicting` protocols and a
   `PredictorFactory` value type. Production code uses the default
   factory which builds the real CoreML-backed predictors ; tests can
   pass a factory that returns canned mocks.
 * `SceneUnderstandingController` now takes the factory at init and
   stores `any SemSegPredicting` / `any DepthPredicting` instead of
   concrete types.

P3.1 — split GardenARPlacementView (3300+ lines)
 * Promoted `GardenARPlacementContainerView` from fileprivate to
   internal so the Coordinator can be reached from extensions in
   other files.
 * Extracted ML wiring to `GardenARPlacementView+SceneUnderstanding.swift`
   (`syncSceneUnderstanding`, `applySceneSnapshot`).
 * Extracted the entire manual-replacement flow (issue #111) to
   `GardenARPlacementView+ManualReplacement.swift` — 15 methods,
   ~430 lines including the morphing + ghost rendering helpers.
 * Main file is now 3418 lines (down from 3803), with the ML and
   manual-replacement code living next to the rest of their feature.

Refs: #188
